// src/types/ts-rs/SpellView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SpellView = { id: bigint, name: string, iconUrl: string | null, level: string, school: string, rangeArea: string, attackSave: string, components: string, castTime: string, duration: string, damageEffect: string, dice: string | null, };